#21 JWS, JWE suite class modify to enum class
