+++ syruppay-jose/src/main/java/com/skplanet/jose/jws/JwsHeader.java
@@ -25,7 +25,7 @@
 package com.skplanet.jose.jws;
 
 import com.skplanet.jose.JoseHeader;
-import com.skplanet.jose.jwa.suits.JwsAlgorithmSuites;
+import com.skplanet.jose.jwa.suites.JwsAlgorithmSuites;
 
 /**
  * Created by byeongchan.park@sk.com(1000808) on 2015-12-22.
